E-commerce-Data-Warehouse-and-Batch-Analytics-Pipeline/requirements.txt
apache-airflow==2.10.4
pandas==2.2.3
psycopg2-binary==2.9.10